+++ src/utils.ts
@@ -18,10 +18,7 @@ export function logDebug(message: string, args: Record<string, any>): void {
  * @param id GTM ID
  * @param params query params object
  */
-export function loadScript(
-  id: string,
-  config: Pick<VueGtmUseOptions, "defer" | "queryParams"> = {}
-): void {
+export function loadScript(id: string, config: Pick<VueGtmUseOptions, "defer" | "queryParams"> = {}): void {
   const win = window,
     doc = document,
     script = doc.createElement("script"),